+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelGenerator.cpp
@@ -249,7 +249,7 @@ ModelGenerator::defineElements(std::vector<Element> &elements, std::map<std::str
 
 NetworkComponent::StateChange_t
 ModelGenerator::evalZ(const double& /*t*/) {
-  State currState = static_cast<State>(z_[0]);
+  State currState = static_cast<State>(static_cast<int>(z_[0]));
   if (currState != getConnected()) {
     Trace::debug() << DYNLog(GeneratorStateChange, id_, getConnected(), z_[0]) << Trace::endline;
     if (currState == OPEN) {
