+++ dynawo/sources/Models/CPP/ModelNetwork/DYNModelHvdcLink.cpp
@@ -194,7 +194,7 @@ ModelHvdcLink::evalJtPrim(SparseMatrix& /*jt*/, const int& /*rowOffset*/) {
 NetworkComponent::StateChange_t
 ModelHvdcLink::evalZ(const double& /*t*/) {
   // evaluation of the discrete variables current values
-  State currState1 = static_cast<State>(z_[0]);
+  State currState1 = static_cast<State>(static_cast<int>(z_[0]));
   if (currState1 != getConnected1()) {
     Trace::debug() << DYNLog(Converter1StateChange, id_, getConnected1(), z_[0]) << Trace::endline;
     if (currState1 == OPEN) {
@@ -208,7 +208,7 @@ ModelHvdcLink::evalZ(const double& /*t*/) {
     stateModified_ = true;
   }
 
-  State currState2 = static_cast<State>(z_[1]);
+  State currState2 = static_cast<State>(static_cast<int>(z_[1]));
   if (currState2 != getConnected2()) {
     Trace::debug() << DYNLog(Converter2StateChange, id_, getConnected2(), z_[1]) << Trace::endline;
     if (currState2 == OPEN) {
